third_party/WebKit/LayoutTests/fast/input/input-device-constructor.html

<!DOCTYPE html>
<html>
<head>
<script src="../../resources/js-test.js"></script>
</head>
<body>
<script>
description("This tests the constructor for the InputDeviceCapabilities DOM class.");

shouldBeFalse("new InputDeviceCapabilities().firesTouchEvents");
shouldBeTrue("new InputDeviceCapabilities({ firesTouchEvents: true }).firesTouchEvents");

</script>
</body>
</html>
